# docs/report_review_productization/openapi-v1.yaml
openapi: 3.1.0
info:
  title: ZQ Report Review Control API
  version: 0.1.0
servers:
  - url: https://demo-domain.invalid/api/v1
paths:
...
  /admin/billing-holds:
    get:
      operationId: uncertainBillingHolds
      security: [{bearerAuth: []}]
      description: Administrator only. Paginated uncertain holds, without provider secrets or request bodies.
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
      responses:
        "200": {description: Items and nullable next_offset}
        "401": {description: Login required}
        "403": {description: Administrator required}
  /admin/billing-holds/{hold_id}/reconciliation:
    parameters:
      - {name: hold_id, in: path, required: true, schema: {type: string}}
    get:
      operationId: getBillingReconciliation
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Administrator-only audit record}
        "401": {description: Login required}
        "403": {description: Administrator required}
        "404": {description: No reconciliation record}
    post:
      operationId: submitBillingReconciliation
      security: [{bearerAuth: []}]
      description: Verified total customer charge for the whole hold, not an additional amount. Identical submission is idempotent. Actor is taken from authentication, never request data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingReconciliationRequest'
      responses:
        "200": {description: Committed or identical existing reconciliation}
        "401": {description: Login required}
        "403": {description: Administrator required}
        "404": {description: Hold missing}
        "409": {description: Conflicting decision or ineligible state}
        "422": {description: Invalid input, rejected values not echoed}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BillingReconciliationRequest:
      type: object
      additionalProperties: false
      required: [confirmed_amount, evidence_sha256, evidence_reference]
      properties:
        confirmed_amount:
          description: Nonnegative finite decimal, at most 8 decimal places, less than 1000000000000. Decimal string recommended to preserve precision.
          type: [number, string]
        evidence_sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        evidence_reference: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'}
...